+++ client/src/components/CocktailDisplay.tsx
@@ -1,13 +1,12 @@
 import { FaComment, FaEllipsisV, FaTrash } from "react-icons/fa";
-import { getUserPicture } from "../utils/functions";
 import IconMenu from "./IconMenu";
-import TimeAgo from "javascript-time-ago";
 import { Link, useNavigate } from "react-router-dom";
 import { Carousel } from "react-responsive-carousel";
-import LikeButton from "./LikeButton"; 
+import LikeButton from "./LikeButton";
 import { authPost } from "../utils/requests";
 import { toast } from "react-toastify";
 import { ICocktail } from "../types/cocktail";
+import ItemUser from "./ItemUser";
 
 interface CocktailDisplayProps {
     cocktail: ICocktail;
@@ -16,7 +15,6 @@ interface CocktailDisplayProps {
 
 const CocktailDisplay = ({ cocktail, likeUnlike }: CocktailDisplayProps) => {
     const user = JSON.parse(localStorage.getItem("user") || "{}");
-    const timeAgo = new TimeAgo("en-US");
 
     const navigate = useNavigate();
     const options = [
@@ -63,36 +61,10 @@ const CocktailDisplay = ({ cocktail, likeUnlike }: CocktailDisplayProps) => {
         navigate(`/cocktail/${cocktail._id}`);
     };
 
-    const gotoUserProfile = (
-        e: React.MouseEvent<HTMLDivElement | HTMLParagraphElement, MouseEvent>
-    ) => {
-        e.stopPropagation();
-        navigate(`/user/${cocktail.user._id}`);
-    };
-
     return (
         <div className="bg-[#212121] rounded-lg border border-zinc-700">
             <div className="p-4 flex items-center">
-                <div
-                    onClick={gotoUserProfile}
-                    className="h-8 w-8 rounded-full overflow-hidden cursor-pointer"
-                >
-                    <img
-                        src={getUserPicture(cocktail.user)}
-                        alt="Avatar"
-                        className="h-full w-full object-cover"
-                    />
-                </div>
-                <div className="ml-3 flex-1">
-                    <Link to={`/user/${cocktail.user._id}`}>
-                        <p className="text-sm font-medium hover:text-[#D93025] hover:underline inline">
-                            {cocktail.user.f_name} {cocktail.user.l_name}
-                        </p>
-                    </Link>
-                    <p className="text-xs text-zinc-400">
-                        {timeAgo.format(new Date(cocktail.createdAt))}
-                    </p>
-                </div>
+                <ItemUser user={cocktail.user} createdAt={cocktail.createdAt} />
                 {user._id === cocktail.user._id && (
                     <IconMenu
                         Icon={<FaEllipsisV className="h-4 w-4" />}
