+++ client/src/components/CommentItem.tsx
@@ -1,9 +1,8 @@
 import React from "react";
 import { IComment } from "../types/comment";
-import { getUserPicture } from "../utils/functions";
 import LikeButton from "./LikeButton";
 import { UserPost } from "../types/user";
-import { useNavigate } from "react-router-dom";
+import ItemUser from "./ItemUser";
 
 interface CommentItemProps {
     comment: IComment;
@@ -28,27 +27,10 @@ const CommentItem: React.FC<CommentItemProps> = ({
         reply(commentId);
     };
 
-    const navigate = useNavigate();
-
-    const handleProfileClick = (userId: string) => {
-        navigate(`/user/${userId}`);
-    };
-
     return (
         <div className="p-4 bg-[#2a2a2a] rounded-xl shadow-md">
-            <div className="flex items-center mb-2">
-                <img
-                    src={getUserPicture(comment.user)}
-                    alt={comment.user._id}
-                    className="w-10 h-10 rounded-full mr-2 cursor-pointer"
-                    onClick={() => handleProfileClick(comment.user._id)}
-                />
-                <div
-                    className="font-semibold text-white hover:text-red-500 hover:underline cursor-pointer"
-                    onClick={() => handleProfileClick(comment.user._id)} 
-                >
-                    {comment.user.f_name} {comment.user.l_name}
-                </div>
+            <div className="p-4 flex items-center">
+                <ItemUser user={comment.user} createdAt={comment.createdAt} />
             </div>
             <p className="text-gray-400 mb-4 ml-8">{comment.content}</p>
             <div className="flex items-center gap-4">
@@ -64,7 +46,7 @@ const CommentItem: React.FC<CommentItemProps> = ({
                 >
                     Reply
                 </button>
-                <span className="text-gray-500" >
+                <span className="text-gray-500">
                     {comment.replies.length} Replies
                 </span>
             </div>
